# src/main/apiServer/generated/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cherry Studio API",
    "version": "1.0.0",
    "description": "OpenAI-compatible API for Cherry Studio with additional Cherry-specific endpoints",
    "contact": {
      "name": "Cherry Studio",
      "url": "https://github.com/CherryHQ/cherry-studio"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use the API key from Cherry Studio settings"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/v1/history/topics": {
      "get": {
        "summary": "List chat history topics",
        "description": "Returns a lightweight topic catalog for local history browsing. Time range filters may be provided either as JSON range query params or as from/to pairs.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "query",
            "name": "topicCreatedFrom",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "topicCreatedTo",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "topicActivityFrom",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "topicActivityTo",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "assistantId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minMessageCount",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["createdAt", "updatedAt", "lastMessageAt", "messageCount"]
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Topic catalog"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "503": {
            "description": "Renderer is not ready"
          }
        }
      }
    },
    "/v1/history/topics/{topicId}": {
      "get": {
        "summary": "Get topic metadata",
        "description": "Returns a single topic metadata record with stable counts and timestamps.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Topic metadata"
          },
          "404": {
            "description": "Topic not found"
          },
          "503": {
            "description": "Renderer is not ready"
          }
        }
      }
    },
    "/v1/history/topics/{topicId}/messages": {
      "get": {
        "summary": "List topic messages",
        "description": "Returns lightweight message records with previews and message annotations.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": ["user", "assistant"]
            }
          },
          {
            "in": "query",
            "name": "segmentId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message list"
          },
          "404": {
            "description": "Topic not found"
          }
        }
      }
    },
    "/v1/history/messages": {
      "get": {
        "summary": "List messages across topics",
        "description": "Returns a paginated cross-topic message stream for a time window or filtered history scan.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "query",
            "name": "messageFrom",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "messageTo",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "assistantId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "topicId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": ["user", "assistant"]
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cross-topic message stream"
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "503": {
            "description": "Renderer is not ready"
          }
        }
      }
    },
    "/v1/history/topics/{topicId}/transcript": {
      "get": {
        "summary": "Get topic transcript",
        "description": "Returns a paginated transcript view for a topic or a single segment.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "path",
            "name": "topicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "segmentId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": ["user", "assistant", "both"]
            }
          },
          {
            "in": "query",
            "name": "responseSelection",
            "schema": {
              "type": "string",
              "enum": ["all", "preferred"]
            }
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limitMessages",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transcript page"
          },
          "404": {
            "description": "Topic not found"
          },
          "503": {
            "description": "Renderer is not ready"
          }
        }
      }
    },
    "/v1/history/messages/batch": {
      "post": {
        "summary": "Batch get messages",
        "description": "Returns full message records for a list of message ids while preserving input order.",
        "tags": ["History"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["messageIds"],
                "properties": {
                  "messageIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch message records"
          },
          "400": {
            "description": "Invalid request body"
          }
        }
      }
    },
    "/v1/history/messages/{messageId}/context": {
      "get": {
        "summary": "Get message context window",
        "description": "Returns the anchor message plus surrounding conversation messages from the same topic.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "before",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "after",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message context window"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
    "/v1/history/messages/{messageId}": {
      "get": {
        "summary": "Get a single message",
        "description": "Returns a full message record including annotations and assistant-side thinking/tool blocks when available.",
        "tags": ["History"],
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message detail"
          },
          "404": {
            "description": "Message not found"
          }
        }
      }
    },
...
  },
...
}